add answer model + swagger annotations

+++ src/main/java/com/finalproject/testgenerator/services/AnswersService.java
@@ -24,7 +24,7 @@ public class AnswersService {
     }
 
     public Answer createAnswer(Answer answer) {
-        answer.setId(repository.count() + 1);
+        answer.setId((int) (repository.count() + 1));
         answer = repository.save(answer);
         return answer;
     }
